remove: avpro dl method
