--- v3/Editor/Resolver/KinelAvProResolver.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 127c943f4894451c92774ab8fa6b799c
-timeCreated: 1735454797
